+++ src/main/java/com/tryit/graphQL/repositories/AuthorRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuthorRepo extends JpaRepository<Author, Double> {
+public interface AuthorRepo extends JpaRepository<Author, Long> {
 }
